Used Popen for calling ocl_info
In windows, it is just too unreliable to properly redirect stdout for
ocl_info. It will not work if the original stdout is a console because
of different CRTs when crossing the dll boundary. There is no known
solution right now, so the only reliable method is to call a NEW python
(YUCK!) and capture stdout using Popen.