don't let null be a valid state, instead, use unset
the problem with 5aec504 is that null values really shouldn't be allowed; they don't trigger set -u's unbound variable error. the problem was that on bash 3.2, setting a local var would implicitly initialize it to null, whereas in more recent versions of bash, the variable would be left as unset. i had been trying to match the behavior of 3.2, but i have now found that you can simply unset the variable to match the behavior of later versions