fix "fatal: Not a valid object name HEAD:."
if a submodule is not in a subdir, then dirpath is '.', but causes 'git ls-tree HEAD:. submod' to fail with "fatal: Not a valid object name HEAD:.". To fix, append a slash (/). It does not appear that either the submodule's name or path can already contain a trailing slash in the .gitmodules file