Revert commits e8c5906b6ae15c9d1d96b452867148e72743ceaf..16918a9c8fd9b853167386695444f75f36079c69 [continued below the fold...]
The purpose of these commits was to unify the concepts of tagging and pushing to the mirror so that only those refs that get pushed can get tagged. A particular sticking point was the concept of whether a ref that gets tagged, but cannot be pushed because it was re-written and is not allowed to be force pushed, should have the tag-to-the-ref that was created by git_tag_main pushed. I thought it was surprising if it were because you explicitly did not want to force push that ref yet those commits (and the tag to the ref end up in the mirror anyway). But andy pointed out that you did want to push the ref, whether or not it was force pushed, so the commits/tag should be pushed, but the ref itself is not. I think this makes sense.

However, if the push/skip regex's aren't kept in sync between calls to git_tag_main and git_push_main, it is possible to push (note, not force push, as in the previous example) tags-to-refs for refs that you did not want to push in the first place. We agree that this seems undesirable.

Another complication is that these regex's are used not only by the main project but all submodules.

After discussion with andy, it was decided that the behavior of this function will be simplified. Instead of allowing regex's for tag, push/skip and force push, simply have a regex for tag (which still defaults to ^refs/heads/.+), push all refs/tags and allow either all refs to be force pushed or not.