Remove GIT_MIRROR_PUSH_REFS_REGEX & GIT_MIRROR_SKIP_REFS_REGEX and change GIT_MIRROR_FORCE_PUSH_REFS_REGEX to be a flag...
This simplification avoids some of the problems with pushing commits to refs (via the tag created by git_tag_main) that were ostensibly skipped. Additionally, it seems imposible to construct all but the most basic regex that applies correctly to the main repo as well as all submodules. Also, excluding commits in the submodule could lead to a problem if the commit tracked by the superproject is missing.

Alternatively, the regex's could be made to only apply to the main repo, but we decided to simplify things first and add back features as/if needed