remove the message that references the original tag for now...
remove the message that references the original tag in the case of a force push. this was harder to figure out than expected and was of limited usefulness. perhaps if a git merge hook could be created that used the tag to warn a developer that a force push occurred then it might be worth the effort